Improve data loading
Filling the selection list can be slow for large data sets. Making the selection list invisible speeds this up slightly; adding a progress indicator slows it down significantly. Users may not realize the need to wait until the list is filled before trying to interact with it. Making the buttons invisible during loading prevents this.